refact: extract duplicated File/Folder types
